git push "Rant 1"
If there's anybody out there that listens to my plight, listen closely.

This team you see here is nothing but a disorganised mess of lost promises, clashes of ego, and untracked progress. Everything you see before you is just a cold, harsh reality that is the insufferability of this team. I never was on at the start, but the lack of communication drove this shit down the rain. I will log more of this as I progress with this garbage.