Porting 101
The first rule in porting is that if it doesn't work, adapt it to something else. Second rule is to clean up garbage, because now is the greatest time to do so. Third rule is to not introduce anything, because that contaminates the sterile environment; don't want to walk into an operating theatre with some conspicuous tool because it looks pretty, otherwise you risk hurting than fixing