CHECK, CHECK, AND CHECK
There's a reason for IDE warnings, and it's to make sure you are not being stupid with Java. I get it, Java is fun, but at the same time it's is frequently abused and sometimes the coders for Minecraft itself aren't usually the greatest at safe methods. By the way, there is no way in hell I'm touching the 'rewrite' branch because this doesn't deserve the mercy of having swept away mistakes. YOU WILL NEVER LEARN IF YOU JUST BRUSH IT OFF AND SAY "YEAH IT IS BAD". OWN UP TO IT AND FIX IT.