I raise a question, did this mod even work
God knows if it did. I'm trying to operate this thing, and as far as I can tell, this thing barely works as it is. The biomes don't generate, everything's still broken, and things don't spawn or crash. Leave it up to the fact that everything was ported, but how the fuck did this thing even operate at all?